create a new ShopifyOrderAPI for order get.
