+++ src/main/java/com/dft/api/shopify/model/order/v202307/ShopifyOrderRoot.java
@@ -1,0 +1,9 @@
+package com.dft.api.shopify.model.order.v202307;
+
+import lombok.Data;
+
+@Data
+public class ShopifyOrderRoot {
+
+    private ShopifyOrder order;
+}
